% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_icons.R
\name{insert_various_rmd}
\alias{insert_various_rmd}
\alias{rmd_icon_fa_download}
\alias{rmd_icon_fa_external_link}
\alias{rmd_icon_fa_youtube_square}
\alias{rmd_icon_fa_youtube}
\alias{rmd_icon_fa_video}
\alias{rmd_icon_fa_file_pdf}
\alias{rmd_icon_fa_file}
\title{Insert various elements}
\usage{
rmd_icon_fa_download()

rmd_icon_fa_external_link()

rmd_icon_fa_youtube_square()

rmd_icon_fa_youtube()

rmd_icon_fa_video()

rmd_icon_fa_file_pdf()

rmd_icon_fa_file()
}
\description{
Insert various elements
}
\concept{R Markdown formatting add-ins}
